Cover Letter Generator
Extract the keywords from Patient Bills like Patient name, Accession Number (ID of the patient), Date of Birth and the cost of treatment and  generate the cover letters for each patient using those keywords automatically. It is designed to process and generate thousands of patient bills and cover letters at a time.